+++ renderer/pages/_app.jsx
@@ -1,7 +1,6 @@
 import React from 'react';
 import { useState, useEffect } from "react";
-import Header from '../components/Header';
-import Loading from "../components/Loading";
+import Loading from "../components/UI/Loading";
 
 import '../styles/globals.css';
